requirements.txt
kafka-python
oracledb
PyYAML
pydantic
kafka-python